requirements.txt
python-telegram-bot==13.15
flask==2.2.3
requests==2.28.2
gunicorn==20.1.0
psycopg2-binary==2.9.5